refactor: move netDownAmount and grossUpAmount to tokenUtils
